+++ p2smi/README.md
@@ -1,0 +1,15 @@
+# CycloPs_v2: A Cyclic Peptide Library Generator
+
+**CycloPs_v2** is a cyclic peptide library generator designed to facilitate the creation and exploration of diverse peptide libraries, including noncanonical and cyclic peptides. This package was released in its current state based on interest generated by the **PeptideCLM model**, as described in a [bioRxiv paper](https://www.biorxiv.org/content/10.1101/2024.08.09.607221v1).
+
+> **Please cite the PeptideCLM paper** (or its published version) if you use this package. A JOSS paper on this work is forthcoming.
+
+## Usage Guide
+In progress...
+
+## Future Work
+Improvements, including modular functions for peptide modifications and additional SMILES generation options, are in progress.
+
+---
+
+Feedback and contributions are welcome as we continue to develop and refine this tool.
