--- test.fa
@@ -1,20 +1,0 @@
->Seq1|SCSC
-ŽHwLMMWTĠRtHQDHLPGYNWMYNMWÌ
->Seq2|SCCT
-WQFWḀYFIAHYPCQPQIKYcCDYĽIyPLṾḺĊDMWNLEDFNdLTQWNịḂAHWšRLFỆWWAWTtTWHYWCPNyGẼLENFTRTiWVTICKSYMdLCẬ
->Seq3|HT
-WrGṐḤTQṠTYEmwPREAlAIRYLIHPQCNLEVIVQḦTRPE
->Seq4|SCCT
-HFḒRaAMCDMKRLKḺDPCsTWqAŁYVDỰVQDẬ
->Seq5|SCSC
-ṊDIRAkHLQNKWFFNSGeṸDMMDVYFĠẒTEiIVKSeINŻKRDMRDNH
->Seq6|SCNT
-êVFEIḤWEDŶWKTKyGMWMTNWrEGỸKWVIĐ
->Seq7|SCCT
-GHAṧPGEHḮCŔIFNSPACQNYLFENKŤCLĿGCTQFfTLPITĐlKFSKDYQQQVRfRMVẬ
->Seq8|SS
-GTHGIQEVyTWQYNDRŰTFVfFHṺWKKGDECHSECFEÌvAMQẄAṄHWcVEŶQGMPGTMḅṎCIqhHFLMTCDICDF
->Seq9|SCCT
-HFLADCTTĀDSDKKARSỐKmNAGaKEEMHECNALCẟQVDDpSnMMSFSCRRDKmỰWĀHČVNľISAaĮTCVQGẬ
->Seq10|SCSC
-aGICHTPẪḤFMHrsVŲsỂkFDEALYNYLDDMẊFENYRRITALWNNATKCGRLVRMṒHTQGaAŶI
